pyproject.toml
[build-system]
requires = ["setuptools>=75", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-tekna"
dynamic = ["version"]
description = "MCP server for Tekna events and news"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Vortiago" }]
keywords = ["tekna", "events", "news", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "mcp[cli]",
    "httpx",
    "python-dotenv",
    "cachetools",
    "beautifulsoup4",
]

[project.scripts]
mcp-tekna = "mcp_tekna.server:main"

[project.urls]
Homepage = "https://github.com/Vortiago/mcp-tekna"
Repository = "https://github.com/Vortiago/mcp-tekna"

[tool.setuptools_scm]
fallback_version = "0.1.0"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.pyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.10"

[tool.poe.tasks]
test = "pytest tests/ -v"
lint = "ruff check ."
format = "ruff format ."
bump-version = "python scripts/bump_version.py"

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-httpx",
    "ruff",
    "pyright",
    "pre-commit",
    "poethepoet",
]
